requirements.txt
fastapi>=0.115
uvicorn>=0.30
pydantic>=2.8
python-dotenv>=1.0
openai>=1.40
scikit-learn>=1.5
numpy>=1.26
pytest>=8.0
